Also support translations on classes
The module now extends itself with its instance methods
so they are available as class and instance methods.

The include/prepend issue is also resolved by changing how
the scope is set so that it is done on first use instead of
during `initialize`. This change removes the need for an
initializer and that makes the include/prepend issue moot.

Also updated the initializer so Cell uses this translator
`Trailblazer::Translation` instead of `Cell::Translation`
and this removes its dependency on ActionView.